Add rancher k3s upgrader to imported k3s cluster
K3s upgrader will now be installed on imported k3s clusters wwhen
their version is set to a newer version for the first time. On
subsequent occasions, the upgrader will be updated if a new
version of the app is available. This will also use the cattle-system
namespace by default instead of the system-upgrade namespace to prevent
Rancher installations from interfering with user installed upgrader.

Co-authored-by: rmweir <ricardo@rancher.com>