Do not require "local://" principal on users
Do enable a more "gitops" friendly flow we'd rather not prefer this
redudany field was required.  If one uses kubectl apply to create a
user with an external principal and then the controller adds
"local://" is modifies the principal list which is hard to reconcile
later on a subsequent apply.