Modify check to grant owner permissions
Problem: On upgraded setups, clusterrolebinding for system account of a cluster
gets deleted.
The builtin cluster-owner and project-owner roles have a new verb to indicate
ownership. On upgraded setups, the cluster and project roleTemplateBindings
get reconciled before the roletemplates are reconciled. Because of which the
clusterRoleTemplateBinding controller compares a role's rules with the non-updated
cluster-owner roleTemplate. Same happens for project-owner role. This leads to the
deletion of clusterrolebindings that bind the user to owner role.

Solution: This commit adds back the check for role names of "cluster-owner" and
"project-owner" roles. So the role template binding controllers will consider
users with these roles as owners and won't delete the clusterrolebindings.