Add validation for public and private access
Prior, a user could create an EKS cluster that did not have public
or private access enabled. This is invalid and inevitably leads
to an error in the eks-operator. This is because with neither
enabled there is no way to communicate with the cluster. Now, an
error will be returned if neither public nor private access is
enabled.