Add labels to link PRTB/CRTB to RB/CRB  by namespace+name
The projectroletemplatebinding and clusterroletemplatebinding UIDs are used as label keys for
clusterrolebindings and rolebindings in the management cluster as follows
* CRTB.UID is the label key for a CRB, CRTB.UID=memberhsip-binding-owner
* CRTB.UID is label key for the RB, CRTB.UID=crtb-in-project-binding-owner
* PRTB.UID  is the label key for a CRB, PRTB.UID=memberhsip-binding-owner
* PRTB.UID is the label key for RB, PRTB.UID=memberhsip-binding-owner
* PRTB.UID is also the label key for RB, PRTB.UID=prtb-in-cluster-binding-owner
With the backup-restore-operator, during restore if any of the PRTB/CRTB was getting created, its UID
will change, and that won't get reflected on the RB/CRB carrying it as labels. So this commit adds
another label where the key is the namespace+name of the PRTB/CRTB instead of UID, as that will remain same.

PRTB/CRTB are also used as label values for clusterrolebindings and rolebindings in user clusters
* CRTB.UID is label value for CRB, authz.cluster.cattle.io/rtb-owner=CRTB.UID
* PRTB.UID is label key for CRB, PRTB.UID=owner-user
This commit adds a label for the downstream cluster CRBs too to have the namespace+name instead of UID
of the PRTB/CRTB as the label value