Update auth check to use cache
Problem:
When listing large amounts of clusters (200) the call takes approx 40
seconds to complete vs 1 second when using kubectl directly

Solution:
The formatter uses a hard API call to check auth on the clusterTemplate
perms for the user. This adds 200ms per cluster when running locally.
Swap this call to use the cached access through the request
AccessControl