Only validate managed fields
Prior, a user could not update a cluster unless the upstream spec
had been populated on the status. The validator would also
attempt to use the upstream spec to check for invalid
configurations such as no nodegroups or private and public access
being disabled. This was a problem because a cluster may not
have an upstream spec because a field needs to be changed in
its spec before an upstream cluster can exist. Using the upstream
spec in the validation layer also added complexity even though
the validations overlapped with eks-operator and the EKS API.