Treat nil and empty subnets as equal
Lasso has been fixed so that passing an empty slice for a field
does not automatically result in nil. Due to this, saving the same
cluster spec as before would result as an empty slice instead of
nil. These are technically not equal and result in an error. Now,
since they are effectively equal they will be treated as such.